requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
